pmem2: move common ndctl functions to ndctl_region_namespace.c
